NB1: Drop prebuilt camera.msm8998.so
The camera HAL which NB1 is using right now is ancient and it needs libshim_gui
updates every major Android version releases. That is actually time-consuming
to adapt the shims on each Android version upgrade.

Switch to OSS camera HAL, it's actually better this way since we don't have to
deal with a lot of trouble just to get it working everytime Google releases a
new Android version.

Change-Id: I85b813104ea7d62e34bae1bac83288e30fa59e7c